New runChecks setup
